+++ src/NuGet.Core/NuGet.Configuration/Utility/SettingsUtility.cs
@@ -17,6 +17,7 @@ namespace NuGet.Configuration
         private const string GlobalPackagesFolderEnvironmentKey = "NUGET_PACKAGES";
         private const string FallbackPackagesFolderEnvironmentKey = "NUGET_FALLBACK_PACKAGES";
         private const string HttpCacheEnvironmentKey = "NUGET_HTTP_CACHE_PATH";
+        private const string PluginsCacheEnvironmentKey = "NUGET_PLUGINS_CACHE_PATH";
         private const string RepositoryPathKey = "repositoryPath";
         public static readonly string DefaultGlobalPackagesFolderPath = "packages" + Path.DirectorySeparatorChar;
 
@@ -238,8 +239,24 @@ namespace NuGet.Configuration
         /// <summary>
         ///  Get plugins cache folder
         /// </summary>
-        
+        public static string GetPluginsCacheFolder()
+        {
+            var path = Environment.GetEnvironmentVariable(PluginsCacheEnvironmentKey);
+            if (!string.IsNullOrEmpty(path))
+            {
+                // Verify the path is absolute
+                VerifyPathIsRooted(PluginsCacheEnvironmentKey, path);
+            }
 
+            if (!string.IsNullOrEmpty(path))
+            {
+                path = path.Replace(Path.AltDirectorySeparatorChar, Path.DirectorySeparatorChar);
+                path = Path.GetFullPath(path);
+                return path;
+            }
+
+            return NuGetEnvironment.GetFolderPath(NuGetFolderPath.HttpCacheDirectory);
+        }
 
         public static IEnumerable<PackageSource> GetEnabledSources(ISettings settings)
         {
