+++ src/NuGet.Core/NuGet.Protocol/Utility/CachingUtility.cs
@@ -1,0 +1,53 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.IO;
+using System.Linq;
+using System.Security.Cryptography;
+using System.Text;
+
+namespace NuGet.Protocol
+{
+    public class CachingUtility
+    {
+        private const int BufferSize = 8192;
+
+        public static string ComputeHash(string value)
+        {
+            var trailing = value.Length > 32 ? value.Substring(value.Length - 32) : value;
+            byte[] hash;
+            using (var sha = SHA1.Create())
+            {
+                hash = sha.ComputeHash(Encoding.UTF8.GetBytes(value));
+            }
+
+            const string hex = "0123456789abcdef";
+            return hash.Aggregate("$" + trailing, (result, ch) => "" + hex[ch / 0x10] + hex[ch % 0x10] + result);
+        }
+
+        public static Stream TryReadCacheFile(TimeSpan maxAge, string cacheFile)
+        {
+            var fileInfo = new FileInfo(cacheFile);
+
+            if (fileInfo.Exists)
+            {
+                var age = DateTime.UtcNow.Subtract(fileInfo.LastWriteTimeUtc);
+                if (age < maxAge)
+                {
+                    var stream = new FileStream(
+                        cacheFile,
+                        FileMode.Open,
+                        FileAccess.Read,
+                        FileShare.Read | FileShare.Delete,
+                        BufferSize,
+                        useAsync: true);
+
+                    return stream;
+                }
+            }
+
+            return null;
+        }
+    }
+}
